Azure/wep-app/requirements.txt
Flask==1.1.1
azure-storage-blob==12.3.0
python-dateutil==2.8.1